Revert "Improve code quality by refactoring various files"
