+++ src/main/java/convobot/gui/DialogBox.java
@@ -20,7 +20,6 @@ import javafx.scene.layout.HBox;
  * the speaker's face and a label containing text from the speaker.
  */
 public class DialogBox extends HBox {
-
     @FXML
     private Label dialog;
     @FXML
