+++ src/main/java/convobot/gui/MainWindow.java
@@ -14,10 +14,7 @@ import javafx.scene.layout.VBox;
  * Controller for MainWindow. Provides the layout for the other controls.
  */
 public class MainWindow extends AnchorPane {
-
-    private final ConvoBot convo = new ConvoBot("./data/tasks.txt");
-    private final Image userImage = new Image(getClass().getResourceAsStream("/images/DaUser.png"));
-    private final Image convoImage = new Image(getClass().getResourceAsStream("/images/DaConvo.png"));
+    private static final ConvoBot convo = new ConvoBot("./data/tasks.txt");
 
     @FXML
     private ScrollPane scrollPane;
@@ -27,6 +24,9 @@ public class MainWindow extends AnchorPane {
     private TextField userInput;
     @FXML
     private Button sendButton;
+
+    private final Image userImage = new Image(getClass().getResourceAsStream("/images/DaUser.png"));
+    private final Image convoImage = new Image(getClass().getResourceAsStream("/images/DaConvo.png"));
 
     @FXML
     public void initialize() {
